Added some inline optimizations and started restructuring ads function to do linear numerical integration rather than Gauss-Kronrod
I think I will need to simply do a linear integration (i.e. taking small steps and evaluating function based on previous evaluation) since we are taking an integral of dE/dx over a range in x, but the dE/dx is a function of E which is a function of x. Gauss-Kronrod requires the function us to be able to evaluate the function independently at the Kronrod points, but that is not possible in the current configuration.